+++ rate-limiter-core/src/main/java/com/example/ratelimiter/core/model/RateLimitResult.java
@@ -2,11 +2,11 @@ package com.example.ratelimiter.core.model;
 
 public class RateLimitResult {
     private final boolean allowed;
-    private final long retryAfter;
+    private final double retryAfter;
     private final int remainingTokens;
     private final int limit;
 
-    public RateLimitResult(boolean allowed, long retryAfter, int remainingTokens, int limit) {
+    public RateLimitResult(boolean allowed, double retryAfter, int remainingTokens, int limit) {
         this.allowed = allowed;
         this.retryAfter = retryAfter;
         this.remainingTokens = remainingTokens;
@@ -14,7 +14,7 @@ public class RateLimitResult {
     }
 
     public boolean isAllowed() { return allowed; }
-    public long getRetryAfter() { return retryAfter; }
+    public double getRetryAfter() { return retryAfter; }
     public int getRemainingTokens() { return remainingTokens; }
     public int getLimit() { return limit; }
 }
